# public/relay-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PBB - Hub Relay Server API",
    "version": "1.1.0",
    "description": "OpenAPI definition for the implemented PBB - Hub Relay Server local and hub-to-hub APIs."
  },
  "servers": [
    {
      "url": "https://relay.pbb.ph",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "LaravelSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "laravel_session"
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN"
      },
      "RelayClientKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Relay-Key"
      },
      "RelayHubKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Relay-Hub-Key"
      },
      "RelayProtocolVersion": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Relay-Protocol-Version"
      }
    },
    "schemas": {
      "HopTraceEntry": {
        "type": "object",
        "required": ["hub_id", "event", "at"],
        "properties": {
          "hub_id": { "type": "string", "description": "HQ numeric hub identifier for the relay that appended this hop trace entry." },
          "event": { "type": "string", "description": "Relay-side event recorded for this hop, such as submitted or received." },
          "at": { "type": "string", "format": "date-time" }
        }
      },
      "LocalMessageSubmitRequest": {
        "type": "object",
        "required": ["source_system", "target_systems", "message_type", "payload"],
        "properties": {
          "source_system": { "type": "string" },
          "target_systems": {
            "type": "array",
            "description": "Client-owned logical audience. Relay resolves transport hops from topology and preserves this list across the journey.",
            "items": { "type": "string" }
          },
          "message_type": { "type": "string" },
          "payload_format": { "type": "string", "enum": ["json", "file", "image", "binary"] },
          "payload_version": { "type": "string" },
          "reference_type": { "type": "string" },
          "reference_id": { "type": "string" },
          "payload": { "type": "object", "additionalProperties": true },
          "tags": { "type": "array", "items": { "type": "string" } },
          "priority": { "type": "string", "enum": ["low", "normal", "high", "urgent"] },
          "correlation_id": { "type": "string" },
          "attachments_count": { "type": "integer" },
          "occurred_at": { "type": "string", "format": "date-time" },
          "created_at": { "type": "string", "format": "date-time" },
          "content_hash": { "type": "string" }
        }
      },
      "RelayHopEnvelope": {
        "type": "object",
        "required": ["relay_id", "origin_hq_hub_id", "source_hub_id", "source_system", "target_hq_hub_id", "target_systems", "message_type", "payload"],
        "properties": {
          "relay_id": { "type": "string" },
          "origin_hq_hub_id": { "type": "string", "description": "HQ numeric hub identifier where the message originated." },
          "source_hub_id": { "type": "string", "description": "HQ numeric hub identifier for the relay sending this hop." },
          "source_system": { "type": "string" },
          "target_hq_hub_id": { "type": "string", "description": "Relay-owned next-hop HQ hub identifier for this transport hop." },
          "target_systems": {
            "type": "array",
            "description": "Client-owned logical audience preserved across hops.",
            "items": { "type": "string" }
          },
          "hop_trace": { "type": "array", "items": { "$ref": "#/components/schemas/HopTraceEntry" } },
          "message_type": { "type": "string" },
          "payload_format": { "type": "string", "enum": ["json", "file", "image", "binary"] },
          "payload_version": { "type": "string" },
          "reference_type": { "type": "string" },
          "reference_id": { "type": "string" },
          "payload": { "type": "object", "additionalProperties": true },
          "tags": { "type": "array", "items": { "type": "string" } },
          "priority": { "type": "string", "enum": ["low", "normal", "high", "urgent"] },
          "correlation_id": { "type": "string" },
          "attachments_count": { "type": "integer" },
          "occurred_at": { "type": "string", "format": "date-time" },
          "created_at": { "type": "string", "format": "date-time" },
          "content_hash": { "type": "string" }
        }
      },
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": { "type": "string" }
            }
          }
        }
      },
...
      "BatchReceiveRequest": {
        "type": "object",
        "required": ["messages"],
        "properties": {
          "messages": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RelayHopEnvelope" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/messages": {
      "post": {
        "tags": ["Messages"],
        "summary": "Submit a relay message",
        "description": "Creates an outbound relay message owned by the authenticated local client. The client supplies source_system, payload, and target_systems[]. Relay injects origin/source hub identity and computes next-hop deliveries from topology.",
        "security": [
          { "RelayClientKey": [] },
          { "RelayProtocolVersion": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LocalMessageSubmitRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Queued" },
          "422": {
            "description": "Validation failure",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationError" } } }
          }
        }
      },
      "get": {
        "tags": ["Messages"],
        "summary": "List local messages",
        "description": "Returns only outbound messages owned by the authenticated local client.",
        "security": [
          { "RelayClientKey": [] },
          { "RelayProtocolVersion": [] }
        ],
        "parameters": [
          { "name": "source_system", "in": "query", "schema": { "type": "string" } },
          { "name": "message_type", "in": "query", "schema": { "type": "string" } },
          { "name": "priority", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Message list" } }
      }
    },
...
    "/api/v1/receive": {
      "post": {
        "tags": ["Hub Receive"],
        "summary": "Receive one message from another relay",
        "description": "Accepts one inbound hop envelope, enforces idempotency by relay_id, appends hop_trace, performs local delivery for matching target_systems[], and forwards to eligible adjacent relays when needed.",
        "security": [
          { "RelayHubKey": [] },
          { "RelayProtocolVersion": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RelayHopEnvelope" }
            }
          }
        },
        "responses": {
          "201": { "description": "Received" },
          "200": { "description": "Duplicate receive" }
        }
      }
    },
    "/api/v1/receive-batch": {
      "post": {
        "tags": ["Hub Receive"],
        "summary": "Receive multiple messages",
        "description": "Accepts multiple hop envelopes in one request.",
        "security": [
          { "RelayHubKey": [] },
          { "RelayProtocolVersion": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BatchReceiveRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Batch processed" } }
      }
    },
...
  }
}